Fix draggable stop method and add initial tests for it.
