+++ docs/public/examples/002-draggable-auto-scroll/index.html
@@ -18,6 +18,6 @@
         </svg>
       </div>
     </div>
-    <script type="module" src="index.20d527f4.js"></script>
+    <script type="module" src="index.6ca09920.js"></script>
   </body>
 </html>
