+++ docs/public/examples/008-draggable-drag-handle/index.html
@@ -17,6 +17,6 @@
         </svg>
       </div>
     </div>
-    <script type="module" src="index.47e8f013.js"></script>
+    <script type="module" src="index.6f411036.js"></script>
   </body>
 </html>
